User-Item table is replaced
